Removed f32

+++ game-core/src/system/camera.rs
@@ -17,10 +17,7 @@ impl<'s> System<'s> for Movement {
     );
 
     fn run(&mut self, (players, cameras, mut transforms): Self::SystemData) {
-        let mut player_translation = Vector2 {
-            x: 0.0 as f32,
-            y: 0.0 as f32,
-        };
+        let mut player_translation = Vector2 { x: 0.0, y: 0.0 };
 
         for (_, transform) in (&players, &mut transforms).join() {
             player_translation = transform.translation.truncate() - Vector2 { x: 256.0, y: 256.0 };
@@ -30,7 +27,7 @@ impl<'s> System<'s> for Movement {
             let camera_translation = transform.translation.truncate();
             let player_direction = player_translation - camera_translation;
             let camera_safe_edge =
-                player_direction.normalize_to((transform.scale.truncate() / 4 as f32).magnitude());
+                player_direction.normalize_to((transform.scale.truncate() / 4.0).magnitude());
 
             if player_direction.magnitude2() > camera_safe_edge.magnitude2() {
                 let camera_shift = player_direction - camera_safe_edge;
